requirements.txt
streamlit
pandas
requests
pydantic
pyyaml
supabase
python-dotenv